salary_pred/requirements.txt
fastapi[standard]==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.49.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
requests==2.32.5
pydantic==2.11.9
joblib==1.3.2
plotly==5.15.0
pyyaml==6.0.1
seaborn==0.13.2